feat: collapse组件扩充
